Changed async task which listens for a gameplay tag addition or removal Removed override which works on the number of times a tag is present to avoid confusion
